set delay as signed

+++ include/rtp/RTPIncomingSource.h
@@ -38,8 +38,8 @@ struct RTPIncomingSource : public RTPSource
 	QWORD   firstReceivedSenderTimestamp;
 
 	QWORD	firstFrameTime;
-	DWORD   frameDelay;
-	DWORD   frameDelayMax;
+	int64_t  frameDelay;
+	int64_t  frameDelayMax;
 	DWORD   frameTime;
 	DWORD   frameTimeMax;
 	int32_t frameCaptureDelay;
